refactor(probe): promote commands/probe.rs to a directory
Splits the 1450-line probe.rs into focused submodules:

- types.rs       — All view structs/enums + ProbeResult, PROBE_SUBCOMMANDS,
                   the DEFAULT_* constants
- render.rs      — Tree flattening, path shortening, byte humanisation
                   (display_path, flatten_tree, annotate, humanize_bytes,
                   into_display_entry)
- shell_init.rs  — The five shell_init_* entry points + their helpers
                   (humanize_us, format_unix_ts, civil_from_days,
                   target_matches_filter, etc.)
- mod.rs         — re-exports the public surface, plus summary,
                   deployment_map, app, and show_data_dir entry points

Tests are co-located with the code they cover. Public API
(commands::probe::ProbeResult, summary, deployment_map, show_data_dir,
app, the shell_init_* family, view types, and the DEFAULT_* constants)
is preserved via re-exports.
